pyproject.toml
[project]
name = "agent-framework-researcher"
version = "0.1.0"
description = "Deep research agent — port of langchain-ai/open_deep_research to microsoft/agent-framework"
readme = "README.md"
authors = [
    { name = "Aymeric Le Caer", email = "alecaer@microsoft.com" }
]
requires-python = ">=3.14"
dependencies = [
    "agent-framework>=1.0.0rc6",
    "rich>=14.3.3",
    "tavily-python>=0.7.23",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]
